Complete album system pre-invite requirements
- Refactor post creation to server/client components with SSR search params
- Add CreatePostForm component with smart navigation and album preselection
- Fix post count calculation in all album query functions
- Enhance AlbumDiscovery with proper error handling and loading states
- Replace alert() with proper UI feedback messages and auto-dismiss
- Improve loading spinners and error recovery throughout
- All posts now require album assignment as designed
